requirements.txt

fastapi
uvicorn
sqlalchemy
cx_Oracle
python-dotenv
pydantic

python-jose[cryptography]
passlib[bcrypt]
pytest
